Non-OIDs are now transmitted via XML, too
git-svn-id: https://svn.viathinksoft.com/svn/oidplus/trunk@83 02e4d621-2042-4998-a0bc-9ccd24201011

+++ includes/classes/OIDplusGui.class.php
@@ -325,7 +325,7 @@ class OIDplusGui {
 				$out['text'] .= '<br><p><a href="?goto=oidplus:edit_ra$'.urlencode($ra_email).'">Edit contact info</a></p>';
 			}
 
-			if (OIDplus::authUtils()::isAdminLoggedIn()) {
+			if (!empty($ra_email) && OIDplus::authUtils()::isAdminLoggedIn()) {
 				if (class_exists("OIDplusPageAdminListRAs")) {
 					$out['text'] .= '<br><p><a href="#" onclick="return deleteRa('.js_escape($ra_email).','.js_escape('oidplus:list_ra').')">Delete this RA</a></p>';
 				} else {
